% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expit.R
\name{expit}
\alias{expit}
\title{Calculate expit}
\usage{
expit(x)
}
\arguments{
\item{x}{a real number}
}
\description{
This function calculates expit of the input.
}
\examples{
expit(5)
}
\keyword{logistic}

