% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_datetime.R
\name{inline_datetime}
\alias{inline_datetime}
\title{Inline hook for use with knitr, especially to avoid coercion of date-times to character}
\usage{
inline_datetime(x)
}
\description{
Inline hook for use with knitr, especially to avoid coercion of date-times to character
}
