% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{Calculate logit}
\usage{
logit(p)
}
\arguments{
\item{p}{a probability between 0 and 1}
}
\description{
This function calculates logit of an inputted probability.
}
\examples{
logit(0.3)
}
\keyword{logistic}

