% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp3alarm.R
\name{mp3alarm}
\alias{mp3alarm}
\title{Plays a musical mp3 as an alarm.}
\usage{
mp3alarm(file = "~/Music/music/us.mp3", time = 15)
}
\arguments{
\item{file}{filename (with path) to an audio file. Seems to require no spaces in path and filename.}

\item{time}{duration of play time, in seconds}
}
\description{
Plays a musical mp3 as an alarm.
}
