% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mp3alarm.R
\name{mp3alarm}
\alias{mp3alarm}
\title{Plays a musical mp3 as an alarm.}
\usage{
mp3alarm(file = "/Users/fjboehm/Music/iTunes/sjb.m4a", time = 5)
}
\arguments{
\item{file}{}

\item{time}{}
}
\description{
Plays a musical mp3 as an alarm.
}

